Jcwd021001 13 (#87)
* fix(auth): fix reset password

* Merge pull request #11 from purwadhikafullstack/JCWD021001-6 (#13)

* feat(initialization): install some package (#15) (#16)


* feat(auth) add navbar and signup page

* feat(auth) add verification page and link it and signup page to back end

* feat(auth) add verification page and link it and signup page to back end

* feat(auth) add verification page and link it and signup page to back end

* fix(auth) fix and add features on auth pages

* fix(auth) fix login

* fix(auth) fix login

* add footer

* (feat)add profile page

* (fix)edit profile page

* (fix)edit profile page

* (fix)edit profile page

* (fix)edit profile page

* (fix)edit profile page

* (feat)add multer

* (fix)upload avatar

* (fix)add user addresses column migration

* (fix) add isMainAddress column

* (fix) remove facebook login

* (fix): siap merge

* (fix) protected route

* (feat) add create address BE

* (feat) add get current location & sync it with opencage

* (fix) fix get current location & sync it with opencage

* (feat) add manage address page

* (feat) add manage address details

* (feat) add change main address & delete address

* (ix) fixing some details

* (fix): fixing some details

* (fix): fixing some details

* (fix): fixing some details

* (fix) fix some changes

* (add) add lat lng column

* (feat) add choose delivery address

* (feat) add find nearest warehouse and shipping cost

* (feat) add find nearest warehouse and shipping cost front end

* (feat) add admin-list

* (feat) admin and warehouse management

* (fix) some responsive design

* (fix) refactor and add validations

* (fix) refactor and bug fixing

---------

+++ apps/api/src/services/userAddress.services.js
@@ -1,3 +1,5 @@
+import UserAddress from '../models/userAddress.model'
+import { Op } from 'sequelize';
 import {
   findUserAddressQuery,
   createUserAddressQuery,
@@ -10,7 +12,8 @@ import {
   removeMainAddressQuery,
   deleteUserAddressQuery,
   findSearchableCityQuery,
-  findSearchableProvinceQuery
+  findSearchableProvinceQuery,
+  findMainAddressQuery
 } from '../queries/userAddress.queries'
 
 export const findUserAddressService = async (id) => {
@@ -77,6 +80,14 @@ export const findSearchableProvinceService = async (name) => {
     }
 }
 
+export const findMainAddressService = async (id) => {
+  try{
+    return await findMainAddressQuery(id)
+  } catch (err){
+    throw err
+  }
+}
+
 export const createUserAddressService = async (
   id,
   specificAddress,
@@ -88,6 +99,13 @@ export const createUserAddressService = async (
   longitude,
 ) => {
   try {
+    const mainAddress = await findMainAddressQuery(id);
+    let isMainAddress = false;
+
+    if (!mainAddress) {
+      isMainAddress = true;
+    }
+
     const res = await createUserAddressQuery(
       id,
       specificAddress,
@@ -97,6 +115,7 @@ export const createUserAddressService = async (
       postalCode,
       latitude,
       longitude,
+      isMainAddress,
     )
     return res
   } catch (err) {
@@ -128,10 +147,40 @@ export const updateMainAddressService = async (id, userId) => {
   }
 }
 
-export const deleteUserAddressService = async (id) => {
+
+export const deleteUserAddressService = async (id, userId) => {
   try {
-    await deleteUserAddressQuery(id)
+    const addressToDelete = await UserAddress.findByPk(id);
+    const addressesCount = await UserAddress.count({
+      where: { userId: userId },
+    });
+
+    if (addressToDelete && addressesCount <= 1) {
+      await deleteUserAddressQuery(id);
+      return; 
+    }
+
+    if (addressToDelete && addressToDelete.isMainAddress) {
+      await deleteUserAddressQuery(id);
+      if (addressesCount > 1) {
+        const anotherAddress = await UserAddress.findOne({
+          where: {
+            userId: userId,
+            id: { [Op.ne]: id }, 
+          },
+          order: [['createdAt', 'ASC']], 
+        });
+
+        if (anotherAddress) {
+          anotherAddress.isMainAddress = true;
+          await anotherAddress.save();
+        }
+      }
+    } else {
+      await deleteUserAddressQuery(id);
+    }
   } catch (err) {
-    throw err
+    throw err;
   }
-}
+};
+
